requirements.txt
# Discord Bot
py-cord>=2.4.0
discord.py>=2.3.0

# Database
asyncpg>=0.28.0
SQLAlchemy>=2.0.0
alembic>=1.11.0

# Web Framework
Flask>=2.3.0
Flask-Session>=0.5.0
aiohttp>=3.8.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
